rework builder
