+++ src/components/Widget/CrudComponent.ts
@@ -42,6 +42,7 @@ export default class CrudComponent extends Vue {
   public showDialog(opts: DialogOpts = {}): void {
     const { widgetProps, mode, listeners } = opts;
     this.activeDialog = createDialog({
+      root: this.$root,
       component: 'WidgetDialog',
       getCrud: () => ({ ...this.crud, closeDialog: this.closeDialog }),
       getProps: () => widgetProps,
@@ -69,6 +70,7 @@ export default class CrudComponent extends Vue {
   public startChangeWidgetTitle(): void {
     const widgetTitle = this.widget.title;
     createDialog({
+      root: this.$root,
       title: 'Change Widget name',
       message: `Choose a new name for '${widgetTitle}'`,
       dark: true,
@@ -84,6 +86,7 @@ export default class CrudComponent extends Vue {
   public startCopyWidget(): void {
     const id = uid();
     createDialog({
+      root: this.$root,
       title: 'Copy widget',
       message: `To which dashboard do you want to copy widget ${this.widget.title}?`,
       dark: true,
@@ -110,6 +113,7 @@ export default class CrudComponent extends Vue {
 
   public startMoveWidget(): void {
     createDialog({
+      root: this.$root,
       title: 'Move widget',
       message: `To which dashboard do you want to move widget ${this.widget.title}?`,
       dark: true,
@@ -144,6 +148,7 @@ export default class CrudComponent extends Vue {
     ].map((opt, idx) => ({ ...opt, value: idx }));
 
     createDialog({
+      root: this.$root,
       title: 'Delete widget',
       message: `How do you want to delete widget ${this.widget.title}?`,
       dark: true,
