+++ src/plugins/spark/features/StepView/StepViewWidget.vue
@@ -21,18 +21,12 @@ export default class StepViewWidget extends WidgetBase {
 <template>
   <component :is="mode" :crud="crud" :class="cardClass" :open-step="openStep">
     <template #toolbar>
-      <WidgetDialogToolbar v-if="inDialog" :crud="crud" :mode.sync="mode">
+      <component :is="toolbarComponent" :crud="crud" :mode.sync="mode">
         <template #actions>
           <ExportAction :crud="crud" />
           <WidgetActions :crud="crud" />
         </template>
-      </WidgetDialogToolbar>
-      <WidgetToolbar v-else :crud="crud" :mode.sync="mode">
-        <template #actions>
-          <ExportAction :crud="crud" />
-          <WidgetActions :crud="crud" />
-        </template>
-      </WidgetToolbar>
+      </component>
     </template>
   </component>
 </template>
