+++ src/styles/quasar.styl
@@ -49,8 +49,8 @@
   height: 100%;
 }
 
-.overflow-none {
-  overflow: none;
+.overflow-unset {
+  overflow: unset;
 }
 
 .overflow-auto {
